# Optimize linearize_cache_indices implementation.