Summary:
Parallelize across threads even more to obtain better performance.

Microbenchmark at following diff shows that:
batch = 1024, T = 50, avg pooling factor = 400
2514us -> 220us (11.4x improvement)

batch = 1024, T = 1000, avg pooling factor = 10
1855us -> 237us (7.8x improvement)

New implementation always wins regardless of whether pooling factor / # of tables are large.

Differential Revision: D34123799